chore: move custom 500 tests to unit
