--- packages/astro/test/fixtures/custom-500-failing/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@test/custom-500-failing",
-  "version": "0.0.0",
-  "private": true,
-  "dependencies": {
-    "astro": "workspace:*"
-  }
-}
