--- packages/astro/test/fixtures/custom-500/astro.config.mjs
@@ -1,4 +1,0 @@
-import { defineConfig } from 'astro/config';
-
-// https://astro.build/config
-export default defineConfig({});
