--- packages/astro/test/fixtures/custom-500/src/pages/index.astro
@@ -1,12 +1,0 @@
----
-throw "some error"
----
-
-<html lang="en">
-<head>
-  <title>Custom 500</title>
-</head>
-<body>
-  <h1>Home</h1>
-</body>
-</html>
